Inline the XML root helper
Emit the shared XML root wrapper once in an existing generated client file instead of creating a dedicated source generator and Go file.

Copilot-Session: a2fed377-7b26-45fc-b3aa-8b6d6df8b46a
